Make updates to fields possible
Includes:
- giving each form control an id that maps to its
  namespace in the JSON data so it can be
  retrieved
- adding some checking of the existing data to see
  if that sent is different (and needs changing)

Implementation of the namespacing has brought up a
few important points.

It's better to return components in a hash for
readability unless their order is important.

Pages' order is variable depending on how the form
flows have been designed.

Fields in a page have a visual order so need to be
in an array.