Add referenced fields to page view
Changes to these effect all instances in the page
so currently grouping them down the bottom.

There's some complexity around ensuring you only
see the fields for data you can actually change.
This tries to deal with that by ignoring the fields
in each page's fieldrefs property which are just
collections of other ones.

Long term it's unsure whether this is the right
approach.